dom/security/test/mixedcontentblocker/test_windowOpen.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Tests for Mixed Content Navigation with window.open</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>

  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>

<body>

<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

let testsCompleted = 0;
const numberOfTestCases = 2;

function markTestCaseComplete() {
  testsCompleted++;

  if (testsCompleted == numberOfTestCases) {
    SimpleTest.finish();
  }
}

window.onmessage = function(event) {
  if (event.data.src.includes("test1")) {
    is(event.data.errorTarget, "https://test1.example.com/tests/dom/security/test/mixedcontentblocker/file_windowOpen.html", "error thrown for failed iframe load should be from test1's iframe.");
    is(event.data.outcome, "blocked", "http iframe should be blocked from loading in child https window.");
    is(event.data.method, "http", "messages from test1 iframe should be HTTP.");
    markTestCaseComplete();
  }
  else if (event.data.src.includes("test2")) {
    is(event.data.triggeringPrincipal, "https://example.com/tests/dom/security/test/mixedcontentblocker/test_windowOpen.html", "triggeringPrincipal for successfully loaded https iframe should be the original test file.");
    is(event.data.outcome, "loaded", "https iframe should be allowed to load in child https window.");
    is(event.data.method, "https", "messages from test2 iframe should be https");
    markTestCaseComplete();
  }
};

function testURLInOpenedWindow(testURL) {
  let openedWindow = window.open("javascript:''","_blank");
  openedWindow.onload = function() {
    openedWindow.document.body.innerHTML = `<iframe id="testframe" src=\"${testURL}\">`

    let httpsIframe = openedWindow.document.getElementById("testframe");

    httpsIframe.onload = function() {
      let triggeringPrincipal = SpecialPowers.wrap(httpsIframe.contentWindow).docShell.currentDocumentChannel.loadInfo.triggeringPrincipal.asciiSpec;
      openedWindow.opener.postMessage({outcome: 'loaded', method: httpsIframe.src.split(":")[0], src: httpsIframe.src, triggeringPrincipal}, 'https://example.com');
      openedWindow.close();
    }
    httpsIframe.onerror = function(error) {
      let errorTarget = error.target.src;
      openedWindow.opener.postMessage({outcome: 'blocked', method: httpsIframe.src.split(":")[0], src: httpsIframe.src, errorTarget}, 'https://example.com');
      openedWindow.close();
    }
  };
};

testURLInOpenedWindow("https://test1.example.com/tests/dom/security/test/mixedcontentblocker/file_windowOpen.html");
testURLInOpenedWindow("https://test2.example.com/tests/dom/security/test/mixedcontentblocker/file_windowOpen.html");

</script>
</body>
</html>
